Complete A-MoreOOP Minimal
