+++ src/data/duke.txt
@@ -1,8 +1,1 @@
-T|false|lol
-D|false|hahaha|2023-10-15T17:00
-D|false|hdeaded|2023-01-23T19:29
-D|false|return book|2023-01-25T14:00
-D|false|go sch|2023-01-23T14:00
-D|false|haha|2023-01-23T14:00
-D|false|dede|2023-01-23T14:50
-D|false|iii|2023-01-23T13:00
+D|false|hahaha|2023-01-23T16:00
