+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 abstract class Task {
     protected final String description;
     protected final boolean isDone;
